Working widget class for minimap Added class for main gamplay widget Reworked widgets for buildings, ships. All logic now in blueprints Reworked factories, now its blueprint function library class
