--- Source/RTSProject/Private/ShipHUD.cpp
@@ -1,10 +1,0 @@
-#include "ShipHUD.h"
-
-#include "RTSPlayerController.h"
-
-void UShipHUD::NativeConstruct()
-{
-	Super::NativeConstruct();
-
-	PlayerController = Cast<ARTSPlayerController>(GetOwningPlayer());
-}
